Some Fixes to use a generic partial update function
